refactor(llm-server): rename ReWoo → Orchestrating and delete ReWoo planner
Three-part cleanup of the ReWoo naming leak tracked in #33515, all as a
single reviewable commit.

- AgentPlannerTypeReWoo → AgentPlannerTypeOrchestrating (string stays "rewoo")
- IsReWOOPlanner() → IsOrchestratingPlanner() (later removed — see part 3)

Persisted "rewoo" string in nb_custom_agents.executor_type is untouched;
Round 3 in #33515 will migrate that.

LlmServerRewooToReact3Enabled is on in every env, so the ReWoo planner
is unreachable at runtime.

Deletions:
- planner_rewoo_2.go (+ _test, _recovery_test)
- planner_rewoo_critiquer.go (+ _test)
- planner_rewoo_plan_critiquer.go
- planner_rewoo_solver.go (+ _test)
- 6 rewoo prompt files
- PromptRewooSolver registry entry
- Prompt* constants + //go:embed lines + switch cases in prompts_repo/svc.go
- LlmServerRewooToReact3Enabled config field, default, post-boot mutation

Config defaults baked in to preserve previous flag-on behavior:
- llm_server_agent_react_max_iterations: 10 → 50
- llm_server_react_critique_enabled: false → true

Helpers moved from the deleted rewoo files to utils.go (still used by
planner_react_2 / planner_react_3):
- extractToolsInvoked
- hasToolFailureMajority

Every read site was guarding an "upgrade to ReAct3" branch that is now
unconditional. Collapsed all sites and deleted the flag.

- executor.go:
  * effective planner type: Orchestrating/ReAct → ReAct3 unconditionally
  * createAgentPlanner: ReAct/ReAct3/Orchestrating all instantiate NewReActAgent3
- executor_planner.go: parallel-exec gate uses only the ReAct3 planner check;
  IsOrchestratingPlanner() removed (no remaining callers)
- 5 orchestrator agent files: dropped `useReact := flag` branches,
  removed the dead ReWoo-format else-branches, dropped ~200 lines of
  awsRewooOutputFormat / awsRewooExamples / gcpRewoo... / azureRewoo...
  package-level constants that no code path can reach
- agent_k8s_debug_hypothesis_e2e_test: removed the hypothesisModeActive()
  gate (always true now)
- agent_finops.go: comment updated
- config.go: LlmServerReAct3Enabled field + default deleted

- AgentPlannerTypeOrchestrating constant — still meaningful intent-name
  for "this agent orchestrates multi-tool work"; the executor unconditionally
  routes it to ReAct3, mirroring the ModelTier pattern where an agent
  declares intent and config decides the implementation.
- planner_react_2.go — file has no runtime callers now, but planner_react_2_test.go
  still tests NBReActPlanner2's structural methods. Deleting the file is a
  separate step (potentially with the tests) that can happen in a follow-up.

- Persisted "rewoo" string in nb_custom_agents.executor_type (needs
  migration + CHECK widen + backfill + dual-read window). Custom agents
  keep working during this window because ExecutorType is metadata only:
  nbCustomAgent.GetPlannerType() always returns Custom, so DB-persisted
  "rewoo" (Orchestrating) never routes through the deleted planner.
- Docs sweep (docs/planner_rewoo.md, docs/caching.md, docs/GLOSSARY.md,
  llm-server CLAUDE.md, llm-server README.md)
- planner_react_2.go deletion if the underlying tests are also retired
- Unused injectKBContext(plannerType) parameter after the skill-lists collapse

Total: ~5900 lines deleted, ~140 added.

Verified: go build / go vet / go test ./... all pass in llm/llm-server.

Fixes #33515
