+++ llm/llm-server/agents/core/memory_extraction_test.go
@@ -1,47 +1,10 @@
 package core
 
 import (
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
-
-func TestReWooSolver_ParseOutput_WithNotebook(t *testing.T) {
-	solver := &ReWooSolver{}
-
-	t.Run("Final answer with notebook update", func(t *testing.T) {
-		output := `
-<final_answer>
-    <thought>Task completed.</thought>
-    <content>I have listed the connections.</content>
-</final_answer>
-<update_notebook>User prefers direct agent usage.</update_notebook>`
-
-		finish, missingInfo, updatedNotebook, err := solver.parseOutput(output, false)
-		assert.NoError(t, err)
-		assert.NotNil(t, finish)
-		assert.Empty(t, missingInfo)
-		assert.Equal(t, "User prefers direct agent usage.", updatedNotebook)
-		assert.Equal(t, "I have listed the connections.", finish.Data)
-	})
-
-	t.Run("Missing info with notebook update", func(t *testing.T) {
-		output := `
-<missing_information>
-    <thought>Need more info.</thought>
-    <required_information>Which environment?</required_information>
-</missing_information>
-<update_notebook>User mentioned 'dev' environment.</update_notebook>`
-
-		finish, missingInfo, updatedNotebook, err := solver.parseOutput(output, false)
-		assert.NoError(t, err)
-		assert.NotNil(t, finish) // Finish is non-nil as it contains Log/Data for the caller
-		assert.Equal(t, "Which environment?", missingInfo)
-		assert.Equal(t, "User mentioned 'dev' environment.", updatedNotebook)
-		assert.Equal(t, "Which environment?", finish.Data)
-	})
-}
 
 func TestRetrieveAndBuildMemoryNotebook_Broadened(t *testing.T) {
 	// Setup mock agent
@@ -54,7 +17,7 @@ func TestRetrieveAndBuildMemoryNotebook_Broadened(t *testing.T) {
 
 		isRetrievalTask := IsDataRetrievalOrActionRequest(req.Query)
 		isInvestigationTask := IsInvestigationRequestTask(req.Query)
-		isSupportedPlanner := mockAgent.GetPlannerType() == AgentPlannerTypeReWoo || mockAgent.GetPlannerType() == AgentPlannerTypeReAct
+		isSupportedPlanner := mockAgent.GetPlannerType() == AgentPlannerTypeOrchestrating || mockAgent.GetPlannerType() == AgentPlannerTypeReAct
 
 		shouldRun := (isInvestigationTask || isRetrievalTask) && isSupportedPlanner
 		assert.True(t, shouldRun)
@@ -65,7 +28,7 @@ func TestRetrieveAndBuildMemoryNotebook_Broadened(t *testing.T) {
 
 		isRetrievalTask := IsDataRetrievalOrActionRequest(req.Query)
 		isInvestigationTask := IsInvestigationRequestTask(req.Query)
-		isSupportedPlanner := mockAgent.GetPlannerType() == AgentPlannerTypeReWoo || mockAgent.GetPlannerType() == AgentPlannerTypeReAct
+		isSupportedPlanner := mockAgent.GetPlannerType() == AgentPlannerTypeOrchestrating || mockAgent.GetPlannerType() == AgentPlannerTypeReAct
 
 		shouldRun := (isInvestigationTask || isRetrievalTask) && isSupportedPlanner
 		assert.True(t, shouldRun)
@@ -76,29 +39,11 @@ func TestRetrieveAndBuildMemoryNotebook_Broadened(t *testing.T) {
 
 		isRetrievalTask := IsDataRetrievalOrActionRequest(req.Query)
 		isInvestigationTask := IsInvestigationRequestTask(req.Query)
-		isSupportedPlanner := mockAgent.GetPlannerType() == AgentPlannerTypeReWoo || mockAgent.GetPlannerType() == AgentPlannerTypeReAct
+		isSupportedPlanner := mockAgent.GetPlannerType() == AgentPlannerTypeOrchestrating || mockAgent.GetPlannerType() == AgentPlannerTypeReAct
 
 		shouldRun := (isInvestigationTask || isRetrievalTask) && isSupportedPlanner
 		assert.False(t, shouldRun)
 	})
-}
-
-func TestReWooPlanner2_Plan_CapturesSolverNotebook(t *testing.T) {
-	planner := &ReWooPlanner2{
-		Notebook: "initial content",
-	}
-
-	// Simulate the logic in planner_rewoo_2.go (merge/append):
-	updatedNotebook := "User preference: use postgres agent."
-	if updatedNotebook != "" {
-		if planner.Notebook != "" && !strings.Contains(updatedNotebook, planner.Notebook) {
-			planner.Notebook += "\n" + updatedNotebook
-		} else {
-			planner.Notebook = updatedNotebook
-		}
-	}
-
-	assert.Equal(t, "initial content\nUser preference: use postgres agent.", planner.Notebook)
 }
 
 func TestIsDataRetrievalOrActionRequest_Extended(t *testing.T) {
